clippy: clear all client warnings
Eight clippy fixes in the client crate, plus one in main.rs that
landed alongside the b148 cards:

- redundant_pattern_matching → is_ok() (counter_tooltip.rs)
- unnecessary_cast → drop i32 → i32 casts (crest.rs, player_stats.rs)
- collapsible_if → fold nested patterns (game_ui/mod.rs)
- bind_instead_of_map → use Option::map (game_ui/mod.rs)
- manual_contains → use slice::contains (kb_cursor.rs)
- new_without_default → derive Default for ShowVerifiedCards (audit.rs)
- doc_lazy_continuation → restructure PlayerCrest docstring (components.rs)
- dead_code → mark unused HUD_BG_INFO and CommandZoneCard.slot as allowed

Workspace now compiles clippy-clean on all three crates.

+++ crabomination_client/src/systems/game_ui/crest.rs
@@ -95,7 +95,7 @@ fn ring_state_for_seat(
 
     let viewer_life = cv.players.iter().find(|p| p.seat == your_seat).map(|p| p.life).unwrap_or(20);
     let threatened = seat == your_seat
-        && (viewer_life <= 5 || lethal_on_board >= viewer_life as i32);
+        && (viewer_life <= 5 || lethal_on_board >= viewer_life);
     if threatened {
         return RingState {
             base: CREST_RING_THREAT_BASE,
